+++ CHANGES.md
@@ -1,3 +1,9 @@
+Version: 1.2
+===============
+<code>new</code> Icons' background now look like in the android 5.0.<br>
+<code>bug fix</code> The inbox style texts display properly.<br>
+<code>bug fix</code> Unnecessary notifications removed.<br>
+<code>improvement</code> Lot of performance and stability improvement.<br>
 Version: 1.1
 ===============
 <code>new</code> Lock the screen with double tap from the quick circle app.<br>